pyproject.toml
[tool.poetry]
name = "silt"
version = "0.1.0"
description = "SILT is a framework for building multiplayer text adventures"
authors = ["Alexandre Gauthier <alex@lab.underwares.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"
pylama = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
